pyproject.toml
[project]
name = "ainiee-cli"
version = "0.1.0"
description = "A CLI-based hard fork of AiNiee for batch translation."
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "pyinstaller",
    "openai>=1.59.3",
    "boto3",
    "cohere",
    "anthropic",
    "google-genai",
    "tiktoken>=0.9.0",
    "numpy>=2.0.2",
    "openpyxl",
    "mediapipe",
    "opencc",
    "beautifulsoup4",
    "lxml",
    "chardet",
    "PyYAML",
    "requests",
    "httpx[http2]",
    "httpx[socks]>=0.27.0",
    "python-pptx",
    "polib",
    "pandas",
    "rich>=13.9.4",
    "tqdm>=4.67.1",
    "jaconv",
    "python-rapidjson",
    "protobuf==4.25.7",
    "EbookLib",
    "python-multipart",
    "natsort",
    "Pillow",
    "language-data",
    "langcodes",
    "msgspec",
    "spacy",
    "sudachipy",
    "sudachidict-core",
    "absl-py",
    "attrs>=19.1.0",
    "flatbuffers>=2.0",
    "bitstring>=4.3.0",
    "configargparse>=1.7",
    "huggingface-hub>=0.27.0",
    "onnx>=1.18.0",
    "onnxruntime>=1.20.1",
    "babeldoc>=0.5.0",
    "orjson>=3.10.14",
    "pdfminer-six==20250416",
    "peewee>=3.17.8",
    "psutil>=7.0.0",
    "pymupdf>=1.25.1",
    "toml>=0.10.2",
    "xsdata[cli,lxml,soap]>=24.12",
    "msgpack>=1.1.0",
    "pydantic>=2.10.6",
    "tenacity>=9.0.0",
    "freetype-py>=2.5.1",
    "python-levenshtein>=0.27.1",
    "rapidocr-onnxruntime>=1.4.4",
    "pyzstd",
    "rtree",
    "scikit-learn>=1.7.1",
    "aiohttp>=3.13.2",
    "aiohttp-cors>=0.8.1",
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "bcrypt>=4.2.1",
    "pyjwt>=2.10.1",
]

[project.scripts]
ainiee = "ainiee_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ModuleFolders", "Resource", "PluginScripts", "StevExtraction", "I18N"]

[tool.hatch.build.targets.sdist]
include = ["ainiee_cli.py", "pyproject.toml", "README.md", "LICENSE"]

